refactor sim tests
